controle-cubo-python/requirement.txt
face-recognition-models==0.3.0
numpy==1.22.3
opencv-python==4.5.5.64
pygame==2.1.2
pyzmq==22.3.0
